AMLS-4752 | Fix for change officer flow bug, removed FurtherUpdates controller, associated view, models and tests
