--- app/models/changeofficer/FurtherUpdates.scala
@@ -1,42 +1,0 @@
-/*
- * Copyright 2019 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package models.changeofficer
-
-import jto.validation.forms.UrlFormEncoded
-import play.api.libs.json.Writes
-
-sealed trait FurtherUpdates
-case object FurtherUpdatesYes extends FurtherUpdates
-case object FurtherUpdatesNo extends FurtherUpdates
-
-object FurtherUpdates {
-  import jto.validation._
-  import utils.MappingUtils.Implicits._
-
-  implicit val formReads: Rule[UrlFormEncoded, FurtherUpdates] = From[UrlFormEncoded] { __ =>
-    import jto.validation.forms.Rules._
-    (__ \ "furtherUpdates").read[Boolean].withMessage("changeofficer.furtherupdates.validationerror") map {
-      case true => FurtherUpdatesYes
-      case false  => FurtherUpdatesNo
-    }
-  }
-
-  implicit val formWrites: Write[FurtherUpdates, UrlFormEncoded] = Write {
-    case FurtherUpdatesYes => "furtherUpdates" -> "true"
-    case FurtherUpdatesNo => "furtherUpdates" -> "false"
-  }
-}
