--- app/views/changeofficer/further_updates.scala.html
@@ -1,62 +1,0 @@
-@*
- * Copyright 2019 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import include._
-@import forms2._
-
-@(f: forms.Form2[_])(implicit request: Request[_],m:Messages)
-
-@header = {
-    @errorSummary(f)
-    @heading(Messages("changeofficer.furtherupdates.heading"), "summary.updateinformation")
-}
-
-@main(
-    title = Messages("changeofficer.furtherupdates.title") + " - " + Messages("summary.updateinformation"),
-    heading = header
-) {
-
-    @form(f, controllers.changeofficer.routes.FurtherUpdatesController.post()) {
-
-        @fieldset(
-            f = f("furtherUpdates-fieldset"),
-            legend = "changeofficer.furtherupdates.title",
-            legendHidden = true,
-            panel = false,
-            classes = Seq("inline")
-        ) {
-
-            @errorGroup(
-                field = f("furtherUpdates")
-            ) {
-                @radio(
-                    f = f("furtherUpdates"),
-                    labelText = "lbl.yes",
-                    value = "true",
-                    inline = true
-                )
-                @radio(
-                    f = f("furtherUpdates"),
-                    labelText = "lbl.no",
-                    value = "false",
-                    inline = true
-                )
-            }
-        }
-
-        @submit(returnLink = false)
-    }
-}
