+++ conf/app.routes
@@ -101,9 +101,6 @@ POST        /update-information/role-in-business                                
 GET         /update-information/new-nominated-officer                                 @controllers.changeofficer.NewOfficerController.get()
 POST        /update-information/new-nominated-officer                                 @controllers.changeofficer.NewOfficerController.post()
 
-GET         /update-information/further-updates                                       @controllers.changeofficer.FurtherUpdatesController.get()
-POST        /update-information/further-updates                                       @controllers.changeofficer.FurtherUpdatesController.post()
-
 GET         /update-information/remove-responsible-person                             @controllers.changeofficer.RemoveResponsiblePersonController.get()
 POST        /update-information/remove-responsible-person                             @controllers.changeofficer.RemoveResponsiblePersonController.post()
